Fix drop_connection for the default connection
While get_connection uses get_default_connection() if there is no specific connection,
drop_connection didn't and therefore did not drop a connection if no specific name
was given.

In `Table::reestablish_connection()` the connection is retrieved like so:

    $connection = $this->class->getStaticPropertyValue('connection',null);

and if no static property 'connection' is set, the `$connection` will be `null`.
This means that currently the default connection is used, but nothing will be
dropped. This fix changes that, so that if the default connection is used,
the default connection is dropped.

An alternative might have been to change this in the call
(`Table::reestablish_connection()` for instance), but it seems to me that keeping
`drop_connection`'s behaviour the same as `get_connection()` is the best way.